load_lambda/requirements.txt
boto3==1.16.4
botocore==1.19.4
jmespath==0.10.0
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.11
psycopg2-binary==2.8.6
